pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "body2colmap"
version = "0.1.0"
description = "Generate multi-view synthetic data for 3D Gaussian Splatting from SAM-3D-Body output"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["3d", "gaussian-splatting", "colmap", "computer-vision", "rendering"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Computer Vision",
]

dependencies = [
    "numpy>=1.20.0",
    "pyrender>=0.1.45",
    "trimesh>=3.9.0",
    "opencv-python>=4.5.0",
    "scipy>=1.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]
yaml = [
    "pyyaml>=5.4.0",
]
viz = [
    "matplotlib>=3.3.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/body2colmap"
Documentation = "https://github.com/yourusername/body2colmap/blob/main/README.md"
Repository = "https://github.com/yourusername/body2colmap.git"
Issues = "https://github.com/yourusername/body2colmap/issues"

[project.scripts]
body2colmap = "body2colmap.cli:main"

[tool.setuptools]
packages = ["body2colmap"]

[tool.setuptools.package-data]
body2colmap = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["body2colmap"]
omit = [
    "*/tests/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "pyrender.*",
    "trimesh.*",
    "cv2.*",
]
ignore_missing_imports = true
